Move KDF to auth
